requirements.txt
aiohttp==3.7.4
Django==3.2.8
django-extensions==3.1.3
django-filter==21.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
drf-yasg==1.20.0
redis==3.5.3
requests==2.26.0
celery==5.1.2
dj_database_url==0.5.0
